requirements.txt
fastmcp
flask
torch
torchvision
matplotlib
pillow
six
